dear-back/dear_back/requirements.txt
dj-database-url==0.5.0
Django==2.2.12
django-cors-headers==3.2.1
django-filter==2.2.0
django-filters==0.2.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.1.0
